Add Bootstrap
I'm still not super fluent in the new Rails default asset pipeline. Webpacker is not the future. Here I used sprokets and importmaps, which are modern-ish, and I could get it working within time constraints.

rough edge: Rails7#new, with the --css=bootstrap option,  defauls to jsbundling-rails and cssbundling-rails, using esbuild. That might be the future/standard practice. However, the rails asset pipeline seems to be a moving target for the past 3 versions, so I'm not going to bother learning the new, new, new way right now due to time constraints.